+++ src/cookiecutter_python/backend/user_config_proxy.py
@@ -1,3 +1,5 @@
+import json
+import logging
 import os
 from typing import Any, MutableMapping
 
@@ -7,6 +9,9 @@ from software_patterns import Proxy, ProxySubject
 from .singleton import Singleton
 
 __all__ = ['get_user_config']
+
+
+logger = logging.getLogger(__name__)
 
 
 my_dir = os.path.dirname(os.path.realpath(__file__))
@@ -20,7 +25,20 @@ class GetUserConfigSubject(ProxySubject[ReturnValueType]):
 
 
 class GetUserConfigProxy(Proxy[ReturnValueType]):
-    pass
+    def request(self, *args, **kwargs):
+        print('\n---- GetUserConfigProxy ----')
+        logger.error(
+            'Get User Config Proxy Request: %s',
+            json.dumps(
+                {
+                    'keyword_args': {k: str(v) for k, v in kwargs.items()},
+                    'positional_args': [str(arg_value) for arg_value in args],
+                },
+                indent=2,
+                sort_keys=True,
+            ),
+        )
+        return super().request(*args, **kwargs)
 
 
 # Singleton and Adapter of Cookiecutter Proxy