+++ src/cookiecutter_python/hooks/pre_gen_project.py
@@ -1,8 +1,14 @@
+import json
 import sys
+from collections import OrderedDict
 
 from cookiecutter_python.backend.input_sanitization import (
     InputValueError,
     build_input_verification,
+)
+from cookiecutter_python.backend.interpreters_support import (
+    InvalidInterpretersError,
+    verify_input_interpreters,
 )
 
 
@@ -12,6 +18,15 @@ def get_request():
     # due to the templated (dynamically injected) code in this file
 
     # the name the client code should use to import the generated package/module
+    print('\n--- Pre Hook Get Request')
+
+    cookiecutter = OrderedDict()
+    cookiecutter: OrderedDict = {{cookiecutter}}
+
+    print('\n', type(cookiecutter['interpreters']))
+    interpreters = cookiecutter['interpreters']
+    if isinstance(interpreters, str):  # we assume it is json
+        interpreters = json.loads(interpreters)
     module_name = '{{ cookiecutter.pkg_name }}'
 
     return type(
@@ -21,6 +36,7 @@ def get_request():
             'module_name': module_name,
             'pypi_package': module_name.replace('_', '-'),
             'package_version_string': '{{ cookiecutter.version }}',
+            'interpreters': interpreters['supported-interpreters'],
         },
     )
 
@@ -50,13 +66,19 @@ def input_sanitization(request):
         raise InputValueError(
             f'ERROR: {request.package_version_string} is not a valid Semantic Version!'
         ) from error
+    try:
+        verify_input_interpreters(request.interpreters)
+    except InvalidInterpretersError as error:
+        # TODO log maybe
+        raise error
+
     print("Sanitized Input Variables :)")
 
 
 def hook_main(request):
     try:
         input_sanitization(request)
-    except InputValueError as error:
+    except (InputValueError, InvalidInterpretersError) as error:
         print(error)
         return 1
     return 0
@@ -64,6 +86,8 @@ def hook_main(request):
 
 def _main():
     request = get_request()
+    # print(request)
+    print('Computed Variables:\n{req}'.format(req=str(request)))
     return hook_main(request)